requirements.txt
imageio~=2.28.0
lazy_loader~=0.2       
numpy~=1.24.3
packaging~=23.1
Pillow~=9.5.0          
pynput~=1.7.6          
pytesseract~=0.3.10    
PyWavelets~=1.4.1      
scikit-image~=0.20.0
scipy~=1.10.1          
six~=1.16.0
tifffile~=2023.4.12